# config/swagger.yaml
swagger: '2.0'
info:
  title: BasePlatform
  description: A NodeJS Platform as a Service
  version: 1.0.0
schemes:
  - http
  - https
basePath: /api/v1
produces:
  - application/json
paths:
...
  '/applications':
    get:
      summary: Application List
      description: |
        Get list of applications
      tags:
        - Applications
      responses:
        '200':
          description: An array of application objects
          schema:
            type: array
            items:
              $ref: '#/definitions/Application'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create Application
      description: |
        Create Application
      parameters:
        - name: title
          in: formData
          description: Title
          required: false
          type: string
        - name: description
          in: formData
          description: Description
          required: false
          type: string
        - name: appPath
          in: formData
          description: Application Path
          required: false
          type: string
        - name: targetServer
          in: formData
          description: Target Server
          required: false
          type: string
        - name: route
          in: formData
          description: Route
          required: false
          type: string
        - name: zip
          in: formData
          description: ZIP File
          required: false
          type: file
      tags:
        - Applications
      responses:
        '200':
          description: A success object
          schema:
            $ref: '#/definitions/Success'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Application:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier
      title:
        type: string
        description: Title
      description:
        type: string
        description: Description
      appPath:
        type: string
        description: App Path
      targetServer:
        type: string
        description: Target Server
      route:
        type: string
        description: Route
...
  Success:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier of object
      type:
        type: string
        description: 'Type of object (user, event, etc)'
      result:
        type: string
        description: Success
      action:
        type: string
        description: 'Create, Update, Delete or [Other]'
      shortMessage:
        type: string
        description: Short message describing success
      longMessage:
        type: string
        description: Long message describing success
      supplementary:
        type: object
        description: Contains supplementary data for the request
  Error:
    type: object
    properties:
      result:
        type: string
        description: Error
      code:
        type: integer
        format: int32
        description: Error code
      shortMessage:
        type: string
        description: Short error message
      longMessage:
        type: string
        description: Long error message
      correlationId:
        type: string
        description: Log ID used for error tracking and troubleshooting